Added PETSc
Added the external matrix-solver library PETSc. The implementation is identical to that of IMAU-ICE. Creation of the sparse matrices all happens in the ice model. The sparse matrix is then handed to the PETSc module, which converts it to the PETSc matrix type, applies the solver, and converts the result back to a regular FORTRAN double-precision array.

While it is generally not significantly faster than the simple SOR solver, the Krylov-solver used by PETSc is much more numerically stable. Especially for high-resolution meshes, where the SOR solver would need a very low relaxation parameter to achieve convergence (thereby slowing down the model to near-uselessness), this greatly improves performance.

Also fixed a few issues with the thermodynamics; the EISMINT1 experiments now work again.