Revamp forcing_module.f90
So it looks as close as possible to its equivalent in IMAU-ICE. For
now, the only truly working option at the moment seems to be
CO2_direct. Inverse routines are WIP. During the big time loop, a
new routine called update_global_forcing is used as a wrapper,
which includes the calls to the subroutines that update the forcing
(insolation, CO2, d18O), those that compute the ice sheet
contributions towards benthic d18O, and the inverse routines.