Replaced CSR matrices with PETSc
Replaced CSR sparse matrices in the remapping code and the mesh operators with PETSc matrices, as the routines I wrote for matrix operations (transpose, multiplication, addition, etc.) are not very efficient.

Note that the velocity solver still uses CSR matrices (as my attempts to convert that to PETSc functionality have not succeeded), which is less-than-optimal, but since matrix assembly and solving is now only a small part of the computation time (most is taken by the secondary DIVA routines), it doesn't affect overall performance much.