Small remapping change
Reintroduced the old CSR-format matrices into the remapping module, to greatly reduce memory use.

When constructing the remapping matrices, the number of non-zeros (both per row and in total) is unknown in advance, and can vary greatly across the rows. Preallocating memory with PETSc is a pain in the ass, but not doing so makes it insanely slow (since apparently when it runs out, it allocates new memory which is only one single element larger, so that it has to copy the arrays every time a single element is added). However, PETSc also has a routine for setting up a PETSc matrix directly from a set of CSR arrays (which are only very slightly differently formatted from the UFEMISM ones). What I did therefore is to use my own CSR routines for setting up the line integral matrices (since my own memory expansion routines are much more efficient), then convert them to PETSc matrices once they're finished, and move on with the PETSc matrix operations from there. This seems to be a good compromise between computational speed and memory use.