Update initialisation of global insolation
Made the subroutine initialise_insolation_data almost identical to
that in IMAU-ICE. The benchmarks checks are now gone, as the
subroutine uses the new standard implemented in IMAU-ICE for this.
This subroutine now calls inquire_insolation_file and
read_insolation_file_time_lat instead of the old versions, which
are identical except for the name. Both versions are included now
in netcdf_module.f90 to ensure compatibility during development of
this temporary branch; old versions will be deleted at the end,
when no longer necessary.

Added extra routine read_insolation_file_timeframes to
netcdf_module.f90, as this will be used when soon introducing the
new subroutine update_insolation_timeframes_from_file.