Add pre-initialisation of regional climate model
Basically all config- and types-related stuff that is needed before
running initialise_model in order to initialise the regional
climate matrix for each model region:

- Added to configuration_module.f90 some climate and SMB config
options.
- Added to data_types_module.f90 the newly named regional climate
and SMB types, plus moved the type_model_region to the bottom so it
can use the new types (temporary, must be repositioned after
cleaning up the new code and removing the old one).
- Added to data_types_netcdf_module.f90 the types needed to read
the climate and SMB data on a regional mesh.