Add initialisation of regional SMB model
Basically all that is needed by initialise_model in order to
initialise the regional SMB model for each model region:

- Added to SMB_module.f90 a revamp of subroutine initialise_SMB_model
following IMAU-ICE's style, accounting for different options for
choice_SMB_model_config. Old version is basically now a subroutine
called initialise_SMB_model_IMAU_ITM, which includes the option to
initialise the firn state from a restart file.
- Added to UFEMISM_main_model.f90 a stop after the initialisation
of the regional SMB model, as the rest of the code still uses the
old version (what remains of it) for the climate stuff.