Update config file for test run on Snellius
Now the default test-simulation runs smoothly on Snellius, including
all domains (ANT, GRL, EAS, NAM) and for at least 1000 years, using
a horizontal resolution of 40--200 km (takes 1.5 min on 32 cores). It
has been also tested with resolutions of 25--200, 20--250 (crashes for
20-200), 15-200, and 10--200 km (latter takes ~50 min on 32 cores).

The execution script for Snellius now specifies the amount of memory
requested, since the default amount per core is not enough when running
very-high-resolution experiments (total >250GB, so the 'thin' partition
could potentially be not enough; change it to 'fat' if needed). The
current script is ok for the small test run though.