Finished initialisation of ocean sub-model
Added the new ocean_module.f90 to the Makefile, which includes all
the subroutines necessary for the initialisation of the new ocean
sub-model. NOTE: It is still missing the implementation of multiple
(ocean) basins for a given domain, used during the extrapolation of
ocean data into ice-shelf cavities.

Added calls to the initialisation subroutines for the ocean sub-model
in both UFEMISM_program.f90 (global stuff ) and UFEMISM_main_model.f90
(regional stuff).