Add subroutine run_ocean_model
Added all necessary stuff to use the new subroutine that runs the
ocean sub-model within the run_model routine. It runs for all the
currently available ocean methods in IMAU-ICE. Keep in mind that
the actual computation of basal melt rates for ice shelves is done
within the BMB_module.f90, and that is not yet implemented.