Add remesh of new ocean submodel
Added the subroutine remap_ocean_model to the ocean module, and a call
to it in UFEMISM_main_model.f90 after a mesh update.

Changed the variable that stores the name of the folder containing the
extrapolated high-res ocean data from local to a permanent field of
the type_ocean_snapshot_regional data type, so it can be used again
during a mesh update.

Minor update to the arguments of the run_BMB_model CALL in
UFEMISM_main_model.f90, to make them compatible with the updated
subroutine.