Reduce full model to minimal version
The model has been reduced to a toy version in order to test
the implementation of the transition from shared memory to
distributed memory. Now the main model program initialises its
configuration, the global forcing, and the 4 domains, the latter
including only the reference geometries, meshes, square grids,
and a few components of the ice model. Then, the model is run
forward in time, only updating the mesh when necessary and
writing output files. There is no computation of ice velocities,
thermodynamics, ice thickness change, GIA, SMB, BMB, etc.

The configuration module (and template file) still has the full
model options. The unused code from the full model is still in
the modules, commented out, except for the UFEMISM_program.f90
and UFEMISM_main_model.f90, and Makefile files.