Add mapping of basins from mesh to ocean grid (ANT+GRL)
The mesh variables containing the definitions for the different basins
(for Antarctica and Greenland) are now mapped onto the high-resolution
grid used for the extrapolation of ocean data under the shelves.
However, the mapping is not efficient for high target regular grid
resolutions (typically 5km, as in BEDMAP2/BEDMACHINE). It works for
Greenland, but stagnants for Antarctica. Need to get a workaround for
this issue, e.g. going from grid to grid ala IMAU-ICE.