Change default ANT basin option from 'file' to 'none'
Set the default preference for Antarctic basins definition in the
configuration module to 'none', as the current code for mapping
these definitions from the mesh to the regular grid needed by the
ocean module (extrapolation of data under the ice shelves) works
but is too slow for such a big domain and high grid resolution. For
this particular job, I will deal with it ala IMAU-ICE, i.e. from
grid-to-grid instead of from mesh-to-grid.