Add placeholder melt-rate extrapolation from FCMP to PMP vertices
Added a routine for extrapolating the melt field from the regular (FCMP)
mask to the (extended) PMP mask. For now, each PMP vertex uses a simple
weighed average of the surrounding FCMP vertices, with the weight depending
on the distance between the PMP and each FCMP, relative to the sum of all
the FCMP distances. Slightly more sophisticated methods (e.g. IMAU-ICE)
coming soon.