Remove old benchmark-exps configuration options
Removed from the configuration module the old IMAU-ICE-style option
do_benchmark_experiment_config, as this was the old-style of running
benchmark experiments. This was replaced by the new style of defining
the geometric, climatic, SMBic, etc. settings in the configuration files,
which was a natural process during the climate porting, but now required
some extra adjustments in the mesh creation, ELRA, and isotopes routines.
The model now does not depend on these options, thus they are now gone.