Add first half of SELEN initialisation
Created the new SELEN modules: main, mapping, and harmonics, which
contain the subroutines used in the first half of the initialisation,
i.e., up until and including the initialisation of the ALF (some data
regarding spherical harmonics of the anchor points) and LONG_TABLE
(some data regarding the entire global SELEN grid).

Added these new modules to the Makefile.

Added some modifications to the UFEMISM_main_model.f90 file, in order
to support some SELEN needs: allocation and initialisation of the
projection parameters within the region%grid_GIA structure, and the
case where C%choice_GIA_model is chosen as "SELEN" during initialisation
of the model region and the main model time loop.

Added to UFEMISM_program a call to initialise SELEN, and to consider
the case where C%choice_sealevel_model is chosen as "SELEN".

Modified a bit the test config file to make testing faster.