Add execution of run_SELEN
Added the sea level equation modules to the Makefile for compilation.
Added a call to run_SELEN in UFE's big time loop.
Added all the subroutines needed in order to run SELEN, to the main
SELEN module: run_SELEN, update_ice_loading_history, etc. Had to
modify a bit the subroutine inverse_sh_transform_single_region_2D,
as it assumes that the GIA grid has the dimensions as (ny,nx), but
in UFE we are transposing everything to the (nx,ny,nz) style.
Added the routine write_to_SELEN_output_file to the NetCDF module,
and the option to output SELEN's deformation rates to the grid and
mesh output files if so wished.
Added the parameter earth_density to the parameters module.