Change name of SELEN field DM -> Dm
When defining the type_SELEN_global, the field DM (used by the sea
level equation module) produces the following error: Function 'tdm'
requires an argument list at (1). It seems that the name DM is already
used by something else (not in UFE) and the compiler complains.
Changed the name of the field to Dm (also changed its accompanying
MPI window to wDm), which solves the problem for now :shrug:.