Add remap of SELEN variables after mesh update
Added the subroutine remap_SELEN_model, which simply reallocates
the fields onto which the SELEN results are mapped after every
call. The "only" (see below) fields that need to be remapped are the bedrock and
geoid deformation rates, which is done in the remap_ice_model
subroutine. NOTE: During the remap of the ice model, the sea level
was simply reallocated, which reset it to 0 and keep it that way until
its next update, i.e. *never* for when using SELEN and only during
model coupling when using the 'eustatic' or 'none' options. That meant
a potential period of time where sea level was wrongly 0, which is
again geen goed idee. Now sea level is always remapped.