Fix SELEN moving window for arbitrary timesteps
There was some old code that prevented the use of an arbitrary time
step for SELEN, particularly values larger than 1000. For very long
simulations (more than a couple glacial cycles) that is a huge pain
in the ass. Modified the way the regular time window is computed and
how it interacts with the irregular one, so now time steps larger or
smaller than 1000 can be used without issue. Still, I would stick to
natural factors of 1000 (e.g. 100, 250, 500. 2000, etc.) to minimise
the risk of errors due to window disalignment. But you do you.