Bypass transect crash for realistic experiments
During the computation of the mesh transects (in the create_transect
routine), if no crossing is found, the program crashes. Since this
computation is not that important for realistic experiments, added a
bypass of this check for them. Not nice to get this error near the
end of a long run. Instead of the crash, now fire a warning. The mesh
is output to a text file for later inspection and bug hunting, as
deadlines permit. Added issue to repo.