Add explicit max res to mesh creation
During the creation of the initial mesh from cart data, the max
resolution was tied to the minimum resolution chosen among all
areas, i.e., the config option was ignored. This created unwanted
high resolution over areas of low interest. It seems that it was
caused by a missing definition of the max res for each submesh.
This is now added to the code, and the initial mesh now looks as
expected.