Add new features to inversion: smoothing, laws, remap
Improved the inversion of basal sliding by allowing the inverted
basal roughness to be remapped after a mesh-update (and after every
iteration, really), so it does not reset to the initial values at any
moment.

Also added a Gaussian smoothing of the inverted field, which does not
decrease the fit of the modelled ice thickness and makes the runs go
much faster (and probably more stable as well).

Added the first lines of code that make the inversion work with any
of the already available sliding laws. So far only the regularised
Coulomb has been testet (GRL and ANT), while the others are also
missing the limits to their respective inverted quantities.

Added to the configuration module a switch to activate the inversion,
so it is controlled from the config files, plus to variables to
control the scaling constant for the inversion and the smoothing
radius.