Import initial geometry from restart file
The model can now fully use an initial geometry read from a restart
file. Although the ability to read initial temperatures from the
same file is already in the model, it is still stored in the region
restart field. Same for the SMB. Those modules need to be adapted
to use the restart data.