Add some low level cosmetic changes
Moved the sea level stuff from the UFEMISM_program and UFEMISM_main
module to a new module called general_sea_level_module,mainly to
avoid having 'explicit' code in the main routines.

Changed the indentation of the main, data types, and configuration
model files to a slightly different standard so it is correctly
folded by most text editors that work with indentantion levels.