Add initialisation of bed roughness from restart file
Now the creation of a restart file includes bed roughness (both
beta_sq and phi_fric; with pure 0s for the not-used one), and the
model is able to initialise bed roughness from this restart file.
This is useful when restarting from a run where the bed roughness
has been inverted, to either continue inverting it or using it as
a temporally constant field.