Update the global scalar output routines
Reordered a bit the output variables, so sea level stuff is shown
first. Also renamed the modelled sea level output to SL_mod, and
added the output of the SL_obs variable.

Now, SL_obs shows the sea level forcing applied to the model,
depending whether it is fixed or prescribed. Fill it with 0s for
eustatic and SELEN modes, of course, as those are not considered a
forcing).

Also removed the old and obsolete memory tracking stuff (a few lines
of code remaining here and there that did nothing significant).