Update restart settings in configuration module
Turns out the restart implementation is more flexible than I thought.

The only thing that will always use the restart data is the mesh. The
rest of restart variables (topography, temperatures, basal stuff ...)
will be stored in the region%restart field and used only if their
respective modes are set to the "restart" option

For example, a simulation can simply load the mesh from a restart
file, while the topography and other variables are initialised as if
it were a fresh run, i.e. from scratch.

Therefore, the config parameters that set the time at which the
restart data will be extracted from the restart file are now located
right beneath the resurrected is_restart parameter, as it is no
longer related exclusively to the topography.