Add remap of restart init topo after mesh update
After a mesh update, the model wants to remap the reference
geometries, which of course includes the init one. If this topo
was read from an init file, it means that there is no data on a grid
from which the usual remap can take the data. Added a modification
that maps the init topo from old to new mesh in this case.