Add calving module (+remove icebergs +remove thin ice)
Added the missing calving stuff, with the basic threshold-thickness
calving method in chain mode; the calving front is checked for points
with not-thick-enough ice and calved out, then again for the new
calving front, then again and again until the entire calving front
has a thickness above the threshold. A new config option sets the max
number of loop rounds to avoid getting stuck on an infinite loop.

Added the removal of floating ice not connected to the grounded ice
sheet (a.k.a fake icebergs), using a mesh-version of the flood-fill
algorithm.

Added the removal of ice that is too thin (less than 1 m), to avoid
these layers of ice everywhere that slow down the model like hell.
I think the mesh generation part could make use of this to avoid
taking into account these artificial layers.