Add modification to MB based on calving front fraction
After introducing the computation of the float_margin_frac variable
to keep track of the effective ice thickness at calving fronts, this
variable is used to scale the mass balance at each calving front
vertex. The idea is that, if the ice front has not advanced enough
to "cover the entire cell", it should not be receiving the mass
balance of the entire cell, but only a percentage according to the
covered fraction.

Tho, it works with neighbouring ice thicknesses rather than areas,
but the idea is more or less the same.