Fix bug during remapping of basal conditions after mesh update
There was a big bug after reallocating the basal roughness variables
for each sliding law in the remap_bed_roughness subroutine, where
the reallocated variables where not reinitialised according to their
expected values ('uniform', 'parameterised', etc.). Fixed this by
remapping the variables when doing an inversion or when using the
'restart' option, and reinitialising for all other cases. now the
model can update the mesh with no issues.