Fix bug in reference topo masks on the regular grid
It seems to me that the masks for land and ocean where flipped, since
grid points where the floating condition was fulfilled were assigned
the "land" type (and "ocean" type elsewhere). This mask is used for
the creation of the initial mesh, but this fix does not bring huge
changes. For some reason the initial mesh creation yields high
resolution in areas where it not asked to do so. We'll see...