Prevent output/remap of IMAU-ITM model when not using it
The variables that are so far exclusive to the IMAU-ITM SMB model
were always assumed to be part of the remapping after a mesh update,
even when not using this particular model. Same with the restart
mode, which did always output the FirnDepth and MeltPreviousYear
variables to the restart file (or expected them during a restart),
even though those are not even allocated when not using IMAU-ITM.
Fixed this by executing these lines of code only when truly needed.