Bring back the choice_ice_dynamics == 'none' option
Some parts of the code gnerated crashes or segmentation faults when
using choice_ice_dynamics == 'none'. This was mainly due to some
velocity-related variables that were always remapped after a mesh
update. Fixed this by initialising the ice velocities pretending that
we are using the DIVA, and the same during the remapping of the
velocitiy variables. In IMAU-ICE, the DIVA is also run once during
initialisation to get a velocity field that can be used to correctly
compute the thermodynamics. This trick is also part of UFEMISM now.