Add tolerance to basal sliding inversion
Now the inversion operates only where the absolute difference in ice
thickness between PD and the model is either above a certain value
(e.g. 100 m) or above a threshold fraction/percentage (e.g. 50% of
the reference ice thickness at that point). Any of those conditions
triggers the inversion to adjust the value of the sliding parameter.