Convert MB values to m of water eq. before outputting
The new routines for outputting the regional scalar variables to a
text file used inconsistent units for SMB/BMB and the rest of the
quantitities (e.g. Melt, Snowfall). Added the conversion of SMB and
BMB to metres of water equivalent per year during the computation of
their equivalent Gt per year.

Also the model uses now Celsius for text output of scalar mean air
temperature over the domain ¯\_(ツ)_/¯