Add inversion of IMAU-ITM melt SMB parameters
Just for fun, and out of curiosity. Now the four IMAU-ITM parameters
can be inverted together or separately during e.g. a steady-state
simulation to get an idea of the best singular values for each of the
domains. The inversion has its own timestep and most of the stuff
(limits, remap, etc.) present in the basal sliding one.