Add reset of BMB_shelf during inversion of melt rates
Added a reinitialisation to the BMB_shelf in 'well' grounded areas
in order to keep the BMB text output consistent with the model.

At each BMB time step, the values of BMB_shelf are reset to 0 if the
vertex is grounded ice in both the model and the reference topo. In,
this case, that vertex is doing well so leave it alone. For points
where the reference is shelf/ocean and the model is grounded (i.e.
grounding line advance) the inverted melt rates are still applied in
the hopes of forcing the GL to go back to its reference point. This
idea also works for other cases: model shelf vs reference grounded
(grodunging line retreat), model shelf vs reference ocean (calving
front advance), etc.