Tweak the inversion of IMAU-ITM parameters
Now the inversion operates only where the parameters are expected to
have an effect: where near-surface air temperatures are at least for
one month of the year above melting (constant and temperature-based
ablation), where the albedo is < 1 (insolation-based ablation), and
where at least one month of the year has near-surface temperatures
below the freezing point (refreezing factor). Still, the constant
ablation factor could be freed from this requirement, as it seems to
provide a better information of the single best value to use over the
entire domain. Or maybe lower the threshold temperature.