Seperate inversion of constant and tempearture ablation
So the temperature-based one operates only where the near-surface
tempearature for any month is above melting, adn the constant
ablation term is free to operate under its own requirement. For now,
that is any point where the above-mentioned temps are less than 5
degrees from the melting point, or above it, of course.