Fix bug: BMB_shelf was resetting to 0 ALL THE TIME
Goddamnit. The supposed resetting of inverted basal melt rates over
'well' behaved grounded vertices (i.e. those that are grounded both
in the model and in the reference topo) was being applied to the
entire field instead. Fixed it. Now the reset operates only on that
particular vertex, as it is supposed to be.