Tweak the simultaneous inversion of SMB parameters
Now the inversion of IMAU-ITM's parameters can safely adjust all its
parameters simultaneously, as each parameter operates on different
areas: Ts only where there are melting temps, Q only where albedo is
less than 1 and Ts is not inverted, and refreezing only where temps
are below melting pointa and there is liquid water. If all parameters
are started from 0 and do not allow for negative numbers, this works
quite well; in this case, refreezing operates where there is a
thickness deficit, Ts operates where there is a superavit and melting
temps, and Q where there is a superavit and no metting temps. During
an inversion, it is better to leave C_abl_constant_inv at 0, as it is
just a threshold, and to invert it only after getting nice single
values for the rest of the parameters.