Increase SMB inversion speed based on weaker dHi_dt requirement
Previously the inversion operated only where the thickness misfit was
not improving at all. This was based on the rate of thickness change
and whether it was different from 0. This meant that even if the
improvement was extremely small, the inversion would skip this point.
Now the SMB inversion operates also on points with "too small" fit
improvements, to speed things up a bit.