Add native computation and output of dHi and dHs
To make life easier regarding plots of the differences between the
modelled and reference PD ice thickness and elevation. New fields
dHi and dHs are now part of the ice model type, and are computed
during the update of the ice thickness at the end of every region
mdoel loop.