Fix order between time step and output
Fixed an issue where the time step for the next region iteration
was being computed *after* the model produced its output files,
causing a weird void that generated inconsistent output time slices.
Now the determination of the next iteration's time step is done before
the output, and the output is done before the region time is updated.
Together with Victor's previous bugfix, now the output is as expected.