Update Jorjo's plotting script for Matlab/Octave
Thoroughly commented the scripts, including the initial section so
the command help works with the plotting function. Integrated (and
removed) the ReadMeshFromFile script into the main function so the
ufe_PlotMeshFile code is a single self contained file/function.

A version translated for Octave is also added, which has almost the
same capabilities as the Matlab one, except for the transparency
stuff, which as of now is not yet implemented for patch objects in
Octave.