Fix restart bug: SMB inversion
When running a restart with the SMB inversion on, the NetCDF routines
were trying to inquire+read the inverted parameters from the previous
simulation even when the SMB initialisation option was set to uniform
(i.e. when we did not want to read the inverted parameters from the
restart file). Fixed this.