Make smoothing of inverted bed friction timestep-independent
Now the inverted basal parameter is contained unmodified in a
separate variable, which is copied to an auxiliary local variable
during smoothing. This way, the smoothing does not accumulate over
time. As a side effect, the previous over-smoothing allowed the
inverted field to "extrapolate" itself over the entire domain; this
is no more, which means an explicit extrapolation is needed now.