Add simplified extrapolation of inverted bed roughness
Since the smoothing of the inverted field is now independent of the
inversion time step, the over-smoothing leaking over the rest of the
domain stopped. To extrapolate over the non-grounded-ice areas of the
domain, a simple extrapolation based on a Gaussian extrapolation on a
mesh is now available in the code. This method is for the time being
simplified in the sense that it only uses the values of the closest
neighbours of the target vertex, as opposed to a specific radius, but
it works alright.

Only values at interior ice sheet points are used to drive the
extrapolation, while the rest of the domain (including grounded ice
margins and groundeds-side grounding lines) receives extrapolated
values.