Bring back SELEN code through C preprocessor
All SELEN-related code was left out of this branch because it could
not be compiled on my Mac M1. Brought it back by using C preprocessor
directives that let the Makefile compile the SELEN code only if so
specified in the Makefile_include file through the variable DO_SELEN.
This way all platforms can have and use the same model code, and the
SELEN stuff can be turned on/off with a single line change.

Since the unselen branch is not necessary anymore, it will get merged
back into develop along all the new stuff in it.