Add separate calving thresholds for sheet/shelf
The calving threshold method was being applied to both grounded and
floating calving fronts, being "too harsh" with grounded marine
margins when using a value of ~100s of meters. Fixed this by adding
a separate ice thickness threshold for ice sheet vertices (which is
for now set to a default of 0 m, i.e. no threshold calving at sheet
fronts).