Fix bug: square grids not always bigger than mesh
For certain square grid resolutions, the margins of the created grid
lied inside the mesh borders, for which the mapping routines did not
have a way to handle it, thus segfault. Fixed this by adding one grid
extra to each side of square grid, to make sure the mesh lies always
completely inside it.