Added ISMIP-style output
Added the option to create ISMIP-compliant output files. When the config flag do_write_ISMIP_output_config = .TRUE., UFEMISM will create a directory in the results folder with the ISMIP-style name. Inside are all the NetCDF files (only on the output grid, not the mesh) for the ISMIP variables, all with the proper names and units.