Add time display + tons of formatting
Prints the current time and time step to screen. At the end of a
coupling interval, prints the average time step for that interval.
Can be turned on/off from the config file. Respects format of output
and mesh creation messages.

Changed a ton of screen message lines, formatting them to look nice
and to be compatible with the new time display.