Fix bug: Wrong stereographic projection subroutines
The subroutines didn't work with reasonable values of alpha_stereo,
which leaked into the scripts for the generation of the reference
topographies. Since the error was in both parts, the model was kinda
consitent, i.e., right for the wrong reasons. This prevented the
direct usage of input data and their corresponding parameters for the
projections. Fixed this by updating the routines based on the latest
version of OBLIMAP (2.0, 2020), which solves this issue. The hard-
coded parameters values in the configuration module were no longer
valid, and needed to be updated to whatever the new input files will
look like, since those also need to be re-generated using correct
projections and parameters.

- Fixed wrong routines in utilities module
- Deleted duplicate routines in mesh_help_functions module
- Redirected all usage of these rutines to the utilites module
- Corrected the configuration module values for the projection