Add global scalar output in NetCDF format
Imported the routines from IMAU-ICE, including the new module
scalar_data_output_module. Output is written to a file at every
coupling interval, but after the model is run for 100 years and
the coupling time is advanced for the next loop. This makes the
prescribed sea level record to be shifted in the output file
(not in the model) by a coupling interval dt. Will fix this.

Old text-based global output still in the model.