Update UFEMISM_program structure
Reordered the way routines are called before and within the
big time loop, to produce correct global output times, and
to clean it up a bit. Now, all updates of forcing fields and
regional/global sea level stuff happen at the start of the
loop; then the output of global variables, and only then the
regional models are run for a coupling interval. Thus, the
running of ice sheets occurs at the end of the loop, right
before updating the coupling time for the next loop. My guess
is that this was not done this way before due to the inverse
d18O and CO2 routines, but those are for now removed from the
model and will wait until their time to be revisited and then
properly implemented.