Fix discrepancy between PD and initial region ice volumes
For the computation of a given region's GMSL contribution, its
volume above floatation (from the mesh) was being compared to the
volume above floatation (from the grid) of the reference topography.
Mapping differences caused the inital model contribution to be not
0, even when starting from exact PD conditions. Fixed this by using
the PD volume from the mesh instead, which is the same as the initial
region volume in these cases.

Moved the computation of a region's area and volume from the main
regional module to the general_sea_level module, since it fits there
better considering that a region's contribution to GMSL is computed
in that routine.