Update GI method to be anomaly compliant
When testing the climate matrix method (GI version) in a future run,
the resulting applied climate was not identical to an equivalent
simulation using constant PD forcing. Reasons for this:
- Bias correction did not yield a corrected climate that was exactly
like the PD climate, due to computer precision limitations. Fixed it
by making the warm GCM snapshot equal to the PD one.
- For precipitation, additionally the EXP(LOG(X)) operation performed
during the climate matrix weighting part had the same issue. Fixed it
by replacing that operation by a simple sum, and adding limiters to
ensure a non-negative precipitation rate.
- The insolation at time=0.0_dp is not the same as what the constant
extrapolation provides when the data record is passed. Fixed this by
enforcing the use of t=0.0_dp for this case. However, this reduces
the flexibility od the model for future runs, so better check this
in the forcing module.