Add new glacial matrix method: precipitation version
Forgot to modify precipitation to be glacial matrix compliant (was
using the enforced glacial index option for my tests, so it worked
for the wrong reasons). Now precipitation approaches the warm snap
perfectly when the CO2 weight approaches 1. This allows for future
projections starting at PD using the climate matrix, something that
is useful for the spin-up protocol.

Brought back the dependency of precipitation on not only geometry
stuff but also on CO2. I have an inkling that I can make it work
as intented in the original paper. We'll see.