Fix some rough ends in glacial matrix
Added a manual switch to use the pure glacial index method if
desired, although as we all know precipitation does not really
work that way for building up a glacial period ice sheet. Anyway,
it is there as an option.

Added a full bias correction to the GCM surface elevation, which
is only used in the computation of lapse rates and orographic
precipitation downscaling for now. Still, surface elevations in
GCM data, specifically for LGM are very likely off compared to
the elevation standard we use in the ice sheet model. WIP.

Renamed a few local variables, added some safety nets, small
bugs, etc.