Improve usage of restart data for inital reference topo
Now when reading the initial geometry from a restart file, a gridded
version is created from it using the PD grid as template (which is
now initialised first). Then the restart data are mapped from the
restart mesh onto this grid. This way, the missing gridded reference
initial topography is brought back to the model, including of course
its secondary data (e.g. masks). This allows to use it after a mesh
update to compute the initial geometry on the new mesh, as usual and
together with the other reference geometries. Much better integration
of the restart feature!