Improve usage of restart data for ice dynamics initialisation
Initialisation of ice variables now uses the regional sea level
from the restart file to compute the initial surface elevations
and thickness above floatation of a model region. Initial masks
are now determined immediately afterwards. In addition, the ice
rates of change (thickness, bedrock, elevation) are also read
from the restart file after the determination of the masks.

Added new config variable initial_guess_sealevel_config, which
can be used during the initialisation of a new simulation that
uses data from a period where the sea level is not necessarily
at present-day level. Valid for the eustatic and SELEN methods
since for the other ones it can be directly prescribed.