Improve numerical stability of model after a restart
During a restart, the ice velocities were initialised to 0.
This caused the velocity solver to take some time to converge,
during which the ice dynamics were not very stable. This in turn
caused significant shocks in the ice volume and overall distribution
in the first years after a restart. Fixed this by introducing the
use of restart velocity data from the previous simulation, plus
performing a model wind-up of ice velocities at the end of the
initialisation, so velocities are not zero at the beginning of
the run. This greatly reduces shocks, at least for simplified
SIA, uniform-everything experiments. Should help a bit at least
during more realistic/sophisticated experiments. Added thermo
to the wind-up for good measure.