Add new memory/senility method for better restarts
Restarts are not perfect, nothing we can do about it for
now. But... here comes a nice tech. If so specified, use
the dHi_dt values from the previous run (which could be the
last time step from a previous run during a restart) and
weigh-average them with the newly computed dHi_dt from
the current time step. The weight depends on how much time
has passed since the start of the run. At the start, full
memory is applied and thus the old dHi_dt is applied as is,
i.e. same rates. As time passes and memory fades, more and
more weight is added to the new dHi_dt, until full senility
kicks in after a specified number of years. At this point
and onwards, no weight is assigned to the original value.
Make sure that total mass is conserved as well, by the way.
This method prevents strong shocks at the beginning of a
restart, plus some other interesting uses. Carpe diem!