Add restart output of averaged dHi_dt and phi_fric
Both variables are computed using a running window, so they
are averages over a user-specified number of ice-dynamics time
steps (for dHi_dt) or bed-roughness-inversion time steps (for
phi_fric). Idea is to use them to improve the stability of the
memory-senility method and a non-inverted run after a calibration,
respectively.