Ensure climate matrix == PD_obs for PD runs
There seems to be lingering artefacts in the way the climate
matrix is initialised and run, which cause that for PD and
future projection runs the climate matrix and PD_obs methods
generate slightly different climates. Fixed this for good.
Works like a charm thanks to the Glacial Matrix.

Took the chance to standardise both the PD_obs and climate
matrix methods way to use the raw, corrected, and applied
fields of climate. Now the "applied" climate contains really
the variables used to compute the fields. For example, the
applied climate elevation is the ice model elevation if any
downscaling is applied to either near-surface temperatures
or precipitation. Any downscaling is done using the climate
variables after any previous correction (e.g. PI bias) as
"source", and the "target" variables are always the applied
fields.